requirements.txt
asn1crypto==1.2.0
cached-property==1.5.1
Cython==0.29.14
python-pkcs11==0.5.0
